Add files via upload
In the database part of the project, it was preferred to use NoSQL instead of using SQL and the project codes were changed according to MongoDb. Again, as shown in the previous update, the ORM structure is also used in the MongoDb database. Mongoose module was used in the project to facilitate MongoDb operations. Added pages to add category option to the system and edit category information.

You can provide database connection by changing the database connection local or global (via mongo db atlas) information through the file in utility / database.js.